+++ test/tracing/database/ioredis/test.js
@@ -295,7 +295,7 @@ describe('tracing/ioredis', function() {
               expect(span.error).to.equal(false);
               expect(span.ec).to.equal(0);
               expect(span.b.s).to.equal(2);
-              expect(span.b.u).to.equal(false);
+              expect(span.b.u).to.not.exist;
               expect(span.data.redis.connection).to.equal(process.env.REDIS);
               expect(span.data.redis.command).to.equal('multi');
               expect(span.data.redis.subCommands).to.deep.equal(['hset', 'hget']);
@@ -332,7 +332,7 @@ describe('tracing/ioredis', function() {
               expect(span.error).to.equal(true);
               expect(span.ec).to.equal(2);
               expect(span.b.s).to.equal(2);
-              expect(span.b.u).to.equal(false);
+              expect(span.b.u).to.not.exist;
               expect(span.data.redis.connection).to.equal(process.env.REDIS);
               expect(span.data.redis.command).to.equal('multi');
               expect(span.data.redis.subCommands).to.deep.equal(['hset', 'hget']);
@@ -367,7 +367,7 @@ describe('tracing/ioredis', function() {
               expect(span.error).to.equal(false);
               expect(span.ec).to.equal(0);
               expect(span.b.s).to.equal(2);
-              expect(span.b.u).to.equal(false);
+              expect(span.b.u).to.not.exist;
               expect(span.data.redis.connection).to.equal(process.env.REDIS);
               expect(span.data.redis.command).to.equal('multi');
               expect(span.data.redis.subCommands).to.deep.equal(['hset', 'hget']);
@@ -414,7 +414,7 @@ describe('tracing/ioredis', function() {
               expect(span.error).to.equal(false);
               expect(span.ec).to.equal(0);
               expect(span.b.s).to.equal(3);
-              expect(span.b.u).to.equal(false);
+              expect(span.b.u).to.not.exist;
               expect(span.data.redis.connection).to.equal(process.env.REDIS);
               expect(span.data.redis.command).to.equal('pipeline');
               expect(span.data.redis.subCommands).to.deep.equal(['hset', 'hset', 'hget']);
@@ -448,7 +448,7 @@ describe('tracing/ioredis', function() {
               expect(span.error).to.equal(true);
               expect(span.ec).to.equal(1);
               expect(span.b.s).to.equal(3);
-              expect(span.b.u).to.equal(false);
+              expect(span.b.u).to.not.exist;
               expect(span.data.redis.connection).to.equal(process.env.REDIS);
               expect(span.data.redis.command).to.equal('pipeline');
               expect(span.data.redis.subCommands).to.deep.equal(['hset', 'hset', 'hget']);
@@ -483,7 +483,7 @@ describe('tracing/ioredis', function() {
               expect(span.error).to.equal(false);
               expect(span.ec).to.equal(0);
               expect(span.b.s).to.equal(2);
-              expect(span.b.u).to.equal(false);
+              expect(span.b.u).to.not.exist;
               expect(span.data.redis.connection).to.equal(process.env.REDIS);
               expect(span.data.redis.command).to.equal('pipeline');
               expect(span.data.redis.subCommands).to.deep.equal(['hset', 'hget']);