+++ test/tracing/database/redis/test.js
@@ -145,7 +145,7 @@ describe('tracing/redis', function() {
               expect(span.error).to.equal(false);
               expect(span.ec).to.equal(0);
               expect(span.b.s).to.equal(2);
-              expect(span.b.u).to.equal(false);
+              expect(span.b.u).to.not.exist;
               expect(span.data.redis.connection).to.equal(process.env.REDIS);
               expect(span.data.redis.command).to.equal('multi');
               expect(span.data.redis.subCommands).to.deep.equal(['hset', 'hget']);
@@ -182,7 +182,7 @@ describe('tracing/redis', function() {
               expect(span.error).to.equal(true);
               expect(span.ec).to.equal(2);
               expect(span.b.s).to.equal(2);
-              expect(span.b.u).to.equal(false);
+              expect(span.b.u).to.not.exist;
               expect(span.data.redis.connection).to.equal(process.env.REDIS);
               expect(span.data.redis.command).to.equal('multi');
               expect(span.data.redis.subCommands).to.deep.equal(['hset', 'hget']);
@@ -219,7 +219,7 @@ describe('tracing/redis', function() {
               expect(span.error).to.equal(true);
               expect(span.ec).to.equal(1);
               expect(span.b.s).to.equal(2);
-              expect(span.b.u).to.equal(false);
+              expect(span.b.u).to.not.exist;
               expect(span.data.redis.connection).to.equal(process.env.REDIS);
               expect(span.data.redis.command).to.equal('pipeline');
               expect(span.data.redis.subCommands).to.deep.equal(['hset', 'hget']);